requirements.txt
apache-airflow==3.1.1
pandas
requests
pyyaml
mlflow
scikit-learn
apache-airflow-providers-standard
apache-airflow-providers-smtp
apache-airflow-providers-fab
fastapi
uvicorn 
prometheus_client 
psutil
joblib 
pydantic 
dvc==3.67.1